+++ proxypay/__init__.py
@@ -2,4 +2,4 @@
 default_app_config = 'proxypay.apps.ProxypayConfig'
 
 # app version
-__version__ = "1.0.3"
+__version__ = "1.1.0"