+++ proxypay/payments/watch.py
@@ -10,9 +10,8 @@ from django.http import HttpResponse
 from django.views.decorators.csrf import csrf_exempt
 
 # proxypay stuffs
-from proxypay.models import Reference
 from proxypay.references import get
-from proxypay.conf import get_private_key, PP_AUTO_PAYMENT_REF_ID
+from proxypay.conf import get_private_key
 
 # ==============================================================================================
 
@@ -39,20 +38,16 @@ def watch_payments(request):
     """View to watch Proxyapy API Webhook"""
 
     if request.method == 'POST':
-        # catch the signature from header
-        signature = request.headers.get('X-Signature')
         # check signature
-        if check_signature(signature, request.body):
+        if check_signature(request.headers.get('X-Signature'), request.body):
             # payment data
             payment = json.loads(request.body)
-            # getting the reference id from custom_fields
-            reference_id = payment.get('custom_fields', {}).get(PP_AUTO_PAYMENT_REF_ID)
             # gettings the referenc by reference id
-            reference = get(reference_id)
+            reference = get(payment.get('reference_id'))
             # chack reference
             if reference:
                 # update as paid
-                reference.paid( payment )
+                reference.paid(payment)
                 # paiment done
                 return HttpResponse(status=200)
             # reference not found